backend/requirements.txt
# backend/requirements.txt

fastapi
uvicorn[standard]
sqlalchemy
psycopg2-binary

# Зависимости для аутентификации
python-jose[cryptography]
passlib[bcrypt]
python-multipart
alembic

# Веб-фреймворк для создания API
fastapi

# ASGI сервер для запуска FastAPI
uvicorn[standard]

# ORM для взаимодействия с PostgreSQL
sqlalchemy

# Драйвер для подключения к PostgreSQL из Python
psycopg2-binary

# Для работы с .env файлом
python-dotenv